deleted the same class
